More progress

+++ include/GUISizing.hpp
@@ -12,7 +12,7 @@ public:
   GUISizing();
   GUISizing(GUIUnit u1, GUIUnit u2, GUIUnit u3);
   GUISizing(GUIUnit u1, GUIUnit u2, GUIUnit u3, GUIUnit offset);
-  void calculate(float root_size, float *l, float *s);
+  void calculate(int root_size, int *l, int *s);
 };
 
 #endif
